Rename DomainTranslator to ExpressionDomainTranslator
Mark ExpressionDomainTranslator as deprecated
