+++ presto-main/src/main/java/com/facebook/presto/sql/planner/EffectivePredicateExtractor.java
@@ -78,9 +78,9 @@ public class EffectivePredicateExtractor
                 return new ComparisonExpression(ComparisonExpression.Operator.EQUAL, reference, expression);
             };
 
-    private final DomainTranslator domainTranslator;
+    private final ExpressionDomainTranslator domainTranslator;
 
-    public EffectivePredicateExtractor(DomainTranslator domainTranslator)
+    public EffectivePredicateExtractor(ExpressionDomainTranslator domainTranslator)
     {
         this.domainTranslator = requireNonNull(domainTranslator, "domainTranslator is null");
     }
@@ -93,9 +93,9 @@ public class EffectivePredicateExtractor
     private static class Visitor
             extends PlanVisitor<Expression, Void>
     {
-        private final DomainTranslator domainTranslator;
+        private final ExpressionDomainTranslator domainTranslator;
 
-        public Visitor(DomainTranslator domainTranslator)
+        public Visitor(ExpressionDomainTranslator domainTranslator)
         {
             this.domainTranslator = requireNonNull(domainTranslator, "domainTranslator is null");
         }
