+++ presto-main/src/main/java/com/facebook/presto/sql/planner/iterative/rule/PickTableLayout.java
@@ -27,7 +27,7 @@ import com.facebook.presto.spi.predicate.NullableValue;
 import com.facebook.presto.spi.predicate.TupleDomain;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.parser.SqlParser;
-import com.facebook.presto.sql.planner.DomainTranslator;
+import com.facebook.presto.sql.planner.ExpressionDomainTranslator;
 import com.facebook.presto.sql.planner.ExpressionInterpreter;
 import com.facebook.presto.sql.planner.LiteralEncoder;
 import com.facebook.presto.sql.planner.LookupSymbolResolver;
@@ -81,13 +81,13 @@ public class PickTableLayout
 {
     private final Metadata metadata;
     private final SqlParser parser;
-    private final DomainTranslator domainTranslator;
+    private final ExpressionDomainTranslator domainTranslator;
 
     public PickTableLayout(Metadata metadata, SqlParser parser)
     {
         this.metadata = requireNonNull(metadata, "metadata is null");
         this.parser = requireNonNull(parser, "parser is null");
-        this.domainTranslator = new DomainTranslator(new LiteralEncoder(metadata.getBlockEncodingSerde()));
+        this.domainTranslator = new ExpressionDomainTranslator(new LiteralEncoder(metadata.getBlockEncodingSerde()));
     }
 
     public Set<Rule<?>> rules()
@@ -113,9 +113,9 @@ public class PickTableLayout
     {
         private final Metadata metadata;
         private final SqlParser parser;
-        private final DomainTranslator domainTranslator;
+        private final ExpressionDomainTranslator domainTranslator;
 
-        private PickTableLayoutForPredicate(Metadata metadata, SqlParser parser, DomainTranslator domainTranslator)
+        private PickTableLayoutForPredicate(Metadata metadata, SqlParser parser, ExpressionDomainTranslator domainTranslator)
         {
             this.metadata = requireNonNull(metadata, "metadata is null");
             this.parser = requireNonNull(parser, "parser is null");
@@ -184,9 +184,9 @@ public class PickTableLayout
     {
         private final Metadata metadata;
         private final SqlParser parser;
-        private final DomainTranslator domainTranslator;
+        private final ExpressionDomainTranslator domainTranslator;
 
-        private PickTableLayoutWithoutPredicate(Metadata metadata, SqlParser parser, DomainTranslator domainTranslator)
+        private PickTableLayoutWithoutPredicate(Metadata metadata, SqlParser parser, ExpressionDomainTranslator domainTranslator)
         {
             this.metadata = requireNonNull(metadata, "metadata is null");
             this.parser = requireNonNull(parser, "parser is null");
@@ -226,7 +226,7 @@ public class PickTableLayout
             PlanNodeIdAllocator idAllocator,
             Metadata metadata,
             SqlParser parser,
-            DomainTranslator domainTranslator)
+            ExpressionDomainTranslator domainTranslator)
     {
         return listTableLayouts(
                 node,
@@ -250,12 +250,12 @@ public class PickTableLayout
             PlanNodeIdAllocator idAllocator,
             Metadata metadata,
             SqlParser parser,
-            DomainTranslator domainTranslator)
+            ExpressionDomainTranslator domainTranslator)
     {
         // don't include non-deterministic predicates
         Expression deterministicPredicate = filterDeterministicConjuncts(predicate);
 
-        DomainTranslator.ExtractionResult decomposedPredicate = DomainTranslator.fromPredicate(
+        ExpressionDomainTranslator.ExtractionResult decomposedPredicate = ExpressionDomainTranslator.fromPredicate(
                 metadata,
                 session,
                 deterministicPredicate,
